Split off module `Primitive.Types.Tx.TxIn.Gen`.
