+++ lib/wallet/test/unit/Cardano/Wallet/Primitive/Types/UTxOSpec.hs
@@ -29,7 +29,7 @@ import Cardano.Wallet.Primitive.Types.TokenPolicy
     ( TokenName (..), TokenPolicyId (..) )
 import Cardano.Wallet.Primitive.Types.Tx
     ( TxIn (..), TxOut (..) )
-import Cardano.Wallet.Primitive.Types.Tx.Gen
+import Cardano.Wallet.Primitive.Types.Tx.TxIn.Gen
     ( coarbitraryTxIn )
 import Cardano.Wallet.Primitive.Types.UTxO
     ( UTxO (..)
